Fix compiler warnings in "Release" mode: unused variables, narrowing, extended initialization (allowed only in c++11), reordering in constructor initialization list, comparison unsigned and signed integers, missing return statement (copy-paste bug), undefined operation (-Wsequence-point).
git-svn-id: https://svnsrv.desy.de/desy/EuXFEL/WP76/karabo/karaboFramework/branches/1.3@17390 a23ab9f9-1ea2-4934-960d-990259f9ea8c

+++ src/karabo/util/Epochstamp.cc
@@ -56,12 +56,12 @@ namespace karabo {
         }
 
         timeval Epochstamp::getTimeOfDay() const {
-            timeval result = {m_seconds, m_fractionalSeconds / 1000000000000ULL};  // std::pow(10, MICROSEC)
+            timeval result = {long(m_seconds), long(m_fractionalSeconds / 1000000000000ULL)};  // std::pow(10, MICROSEC)
             return result;
         }
 
         timespec Epochstamp::getClockTime() const {
-            timespec result = {m_seconds, m_fractionalSeconds / 1000000000ULL};    // std::pow(10, NANOSEC)
+            timespec result = {long(m_seconds), long(m_fractionalSeconds / 1000000000ULL)};    // std::pow(10, NANOSEC)
             return result;
         }
